piperoni-identity-service-main/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
bidict==0.22.1
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.0.1
click==8.1.7
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.7
dulwich==0.21.6
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.17.3
keyring==24.2.0
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
passlib==1.7.4
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
psycopg==3.1.12
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python-engineio==4.8.0
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.1
rapidfuzz==2.15.2
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.9.19
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.5
watchfiles==0.20.0
websockets==11.0.3
wsproto==1.2.0
xattr==0.10.1
yarl==1.9.2
zipp==3.17.0